Crate game

This crate contains everything one needs to represent the backend of the game The Bottom Line, a card game in which players try to increase the value of their company.

Modules

cards

This file contains all information to be able to load in boardgame.json, which is a json representation of the game assets of The Bottom Line.

errors

This file contains all errors that are used by the game, leveraging thiserror. This includes a general GameError enum as well as smaller, action-specific errors used throughout the game logic.

game

This is where the game logic, excluding the player-specific logic, is located.

player

This file contains all four player states, as well as functionality for those players and ways to interact with them.

utility

This file contains utility code used throughout the project.

Statics

SHARED_TS_DIR

The folder containing all shared typescript types.