game

Module cards

This file contains all information to be able to load in boardgame.json, which is a json representation of the game assets of The Bottom Line.

Example:

{
  "metadata": {
    "version": "0.1",
    "gamemode": "board_version"
  },
  "deck_list": {
    "asset_deck": {
      "card_image_back_url": "asset_back.webp",
      "card_list": [
        {
          "title": "R&D Lab",
          "color": "Purple",
          "gold_value": 3,
          "silver_value": 1,
          "copies": 1,
          "ability": "At the end of the game, for one color, turn - into 0 or 0 into +",
          "card_image_url": "assets/rndLab_3-1.webp"
        },
      ]
    },
    "liability_deck": {
      "card_image_back_url": "liability_back.webp",
      "card_list": [
        {
          "liability_type": "Trade Credit",
          "gold_value": 1,
          "copies": 4,
          "card_image_url": "liabilities/tradeCredit_1.webp"
        }
      ]
    },
    "market_events_deck": {
      "card_image_back_url": "market_back.webp",
      "card_list": [
        {
          "title": "Global Treaty on Climate Change",
          "event": {
            "description": "Negotiations have succeeded and all UN-member states have agreed on a global treaty to fight climate change.",
            "effect": "The CSO has to skip their turn to formulate the corporate alignment strategy. All Green assets gain one gold."
          },
          "copies": 1,
          "card_image_url": "events/climateChange.webp"
        },
        {
          "title": "Stable Market",
          "market_status": {
            "rfr": 4,
            "mrp": 4
          },
          "copies": 1,
          "card_image_url": "events/stable_01.webp"
        }
      ]
    }
  }
}

Structs

AssetCard 🔒

Representation of an asset in the json

DeckList 🔒

The list of decks in the json: assets, liabilities and markets/events.

EventCard 🔒

GameData

Overarching struct which contains the asset deck, the liability deck and the market/event deck. This can be used by the game or anything else that might want to get all cards.

LiabilityCard 🔒

Representation of a liability card as it appears in the json

LoadedCards 🔒

Represents the json in its entirety

LoadedCardsMetadata 🔒

Card metadata

MarketEventCard 🔒

Representation of either a market or event card as it appears in the json

MarketStatusCard 🔒

Card representing the non-shared fields with event as it appears in the json

Enums

DataParseError

Errors that can occur when parsing or loading data.

MarketEventDetails 🔒

Enum representing the fields only found on either a market or an event