game

Module common

Re-exports

pub use self::route_sketcher::RouteSketcher;

pub use self::select::RoadSelector;

pub use self::warp::inner_warp_to_id;

pub use self::warp::warp_to_id;

pub use self::warp::Warping;

Modules

route_sketcher 🔒

select 🔒

share

warp 🔒

Structs

CommonState

Functions

cmp_duration_shorter

color_for_agent_type

color_for_trip_phase

jump_to_time_upon_startup

If you want a simulation to start after midnight, pass the result of this to SandboxMode::async_new. It’s less visually overwhelming and more performant to continue with a loading screen than launching the jump-to-time UI. After a deadline of 0.5s, we will switch over to the jump-to-time UI so the user can cancel.

list_names

sync_abtest 🔒

tool_panel