game::debug
Module routes

Structs

AllRoutesExplorer 🔒

See how live-tuned routing parameters affect all requests for the current scenario.

PathCostDebugger

Evaluate why an alternative path wasn’t chosen, by showing the cost to reach every road from one start.

RouteExplorer

See how live-tuned routing parameters affect a single request.

Functions

calculate_demand 🔒

controls_to_params 🔒

params_to_controls 🔒