Module game::devtools::fifteen_min

This is a tool to experiment with the concept of 15-minute neighborhoods. Can you access your daily needs (like groceries, a cafe, a library) within a 15-minute walk, bike ride, or public transit ride of your home?

See https://github.com/dabreegster/abstreet/issues/393 for more context.

Modules

isochrone

Structs

HoverOnBuilding

Viewer
This is the UI state for exploring the isochrone/walkshed from a single building.

Functions

build_panel

draw_star
Draw a star on the start building.