Module game::edit

Re-exports

pub use self::roads::RoadEditor;

pub use self::routes::RouteEditor;

pub use self::stop_signs::StopSignEditor;

pub use self::traffic_signals::TrafficSignalEditor;

pub use self::validate::check_blackholes;

pub use self::validate::check_sidewalk_connectivity;

Modules

crosswalks ğŸ”’

multiple_roads ğŸ”’

After a single road has been edited, these states let the changes be copied to all similar road segments. Note that only lane configuration is copied, not speed limit or access restrictions.

roads ğŸ”’

routes ğŸ”’

stop_signs ğŸ”’

traffic_signals ğŸ”’

validate ğŸ”’

zones ğŸ”’

Structs

ConfirmDiscard

EditMode

LoadEdits

SaveEdits

Functions

apply_map_edits

can_edit_lane

cmd_to_id ğŸ”’

make_changelist ğŸ”’

make_topcenter ğŸ”’

maybe_edit_intersection

speed_limit_choices