game

Module errors

This file contains all errors that are used by the game, leveraging thiserror. This includes a general GameError enum as well as smaller, action-specific errors used throughout the game logic.

Enums

AssetAbilityError

Errors that can happen while performing card abilities

BankerTargetSelectError

Errors related to selecting assets or liabilities when paying off the banker.

DivestAssetError

Errors related to divesting assets.

DrawCardError

Errors related to drawing cards.

FireCharacterError

Errors related to firing a character.

GameError

The main error enum used by the game logic.

GetBonusCashError

Errors related to getting bonus gold

GiveBackCardError

Errors that can happen when a player must give back a card.

LobbyError

Errors that can happen in the lobby phase.

PayBankerError

Errors related to paying the banker on the targets turn

PlayCardError

Errors that can happen when someone plays a card.

RedeemLiabilityError

Errors that can happen when redeeming a liability.

SelectingCharactersError

Errors that can happen while selecting characters.

SwapError

Errors related to swapping hands/cards.

TerminateCreditCharacterError

Errors related to terminating a character’s credit line.