game

Module game

This is where the game logic, excluding the player-specific logic, is located.

Modules

banker_target 🔒

File containing the round state of the game.

lobby 🔒

File containing the lobby state of the game.

results 🔒

File containing the results state of the game.

round 🔒

File containing the round state of the game.

selecting_characters 🔒

File containing the selecting characters state of the game.

Structs

AssetLiabilityCount

Used to return the amount of assets and liabilities that were returned to the deck when the Character::Regulator swaps with the deck.

BankerTargetRound

State containing all information related to the banker targeting state of the game. In the banker target stage, the player that was targeted can elect to issue liabilities and sell off assets at market value in order to raise cash to pay off the banker. Once they have paid off the banker, the game moves to a Results state.

Deck

A wrapper struct around Vec<T> which allows for easy interaction with it as a deck of cards.

Event

The event card type

HandsAfterSwap

Used to return the new hands for the regulator and its player target.

IssuedLiabilityToPayBanker

Struct that represents a liability that a player has selected to be issued to pay off their obligation to the banker. It contains the index of the liability in the hand of the player, as well as the liability itself.

Lobby

State containing all information related to the lobby stage of the game. In the lobby state, players are allowed to join and leave freely. When between 4 to 7 players are in the lobby, players are allowed to start a game.

Market

The market card type

MarketChange

Data used when someone buys a new asset and a market change is triggered

ObtainingCharacters

Used for keeping track of selectable characters in the selecting characters phase.

PickableCharacters

Contains information when picking cards. One gets a list of pickable characters as well as a possible closed character if the player requesting it is the chairman.

PlayerPlayedCard

Data used when someone plays a card

PlayerScore

Representation of a player’s final score, which contains their id as well as their score.

Players

Wrapper struct around Vec<P> to make interacting with them as players internally much easier.

Results

State containing all information related to the results state of the game. In the resuts stage, players can see their scores.

Round

State containing all information related to the round state of the game. In the round stage, players each play a turn where they can draw cards, play cards and use their character ability. After every player has played a turn, players will be able to select characters again. If one player reached six or more assets during a round, the game will move to Results instead.

SelectedAssetsAndLiabilities

A collection of selected assets that will be sold and a list of liabilities that will be issued in order to comply with the banker’s obligation.

SelectingCharacters

State containing all information related to the selecting characters state of the game. In the selecting characters stage, players select a character one by one until everyone has selected a character, after which a round starts.

SoldAssetToPayBanker

Structure that represents an asset that is set to be sold to pay off their obligation to the banker. It contains the index of the asset as well as the market value of the asset.

TurnEnded

Data used when a turn ends

Enums

GameState

The core state representation of The Bottom Line. It has four internal states:

MarketCondition

A representation of the market condition for a specific color. It can either be

Constants

ASSETS_FOR_END_OF_GAME

Amount of assets required to end the game

STARTING_GOLD

Cash each player starts with

Functions

default_backup_deck 🔒