game::game

Enum GameState

pub enum GameState {
    Lobby(Lobby),
    SelectingCharacters(SelectingCharacters),
    Round(Round),
    BankerTarget(BankerTargetRound),
    Results(Results),
}

The core state representation of The Bottom Line. It has four internal states:

Lobby (Lobby)

Selecting Characters (SelectingCharacters)

Round (Round)

Results (Results)

Variants

Lobby(Lobby)

Lobby state of the game. In this state players can freely join and leave the game

SelectingCharacters(SelectingCharacters)

Selecting characters state of the game. In this state each player can select a character, after which the game starts

Round(Round)

Round state of the game. In this state each player plays their turn, which includes drawing cards, playing assets and liabilities and using their character ability

BankerTarget(BankerTargetRound)

Banker target state of the game. In this state a player can do all their actions to pay the baker

Results(Results)

Results state of the game. In this state players can see how they did compared to everyone else

Implementations

impl GameState

pub fn new() -> Self

Creates a new instance of the game. The game starts in lobby state.

Examples

let game = GameState::new();
assert_eq!(game, GameState::Lobby(Lobby::default()));

pub fn lobby(&self) -> Result<&Lobby, GameError>

Tries to get a &Lobby state. Returns an error if the game is not in a lobby state.

Examples

let game = GameState::Lobby(Lobby::default());
assert_eq!(game.lobby(), Ok(&Lobby::default()));

pub fn lobby_mut(&mut self) -> Result<&mut Lobby, GameError>

Tries to get a &mutLobby state. Returns an error if the game is not in a lobby state.

Examples

let mut game = GameState::Lobby(Lobby::default());
assert_eq!(game.lobby_mut(), Ok(&mut Lobby::default()));

pub fn selecting_characters(&self) -> Result<&SelectingCharacters, GameError>

Tries to get a &SelectingCharacters state. Returns an error if the game is not in a selecting characters state.

Examples

let game = GameState::Lobby(Lobby::default());
assert_eq!(game.selecting_characters(), Err(GameError::NotSelectingCharactersState));

pub fn selecting_characters_mut( &mut self, ) -> Result<&mut SelectingCharacters, GameError>

Tries to get a &mutSelectingCharacters state. Returns an error if the game is not in a selecting characters state.

Examples

let mut game = GameState::Lobby(Lobby::default());
assert_eq!(game.selecting_characters_mut(), Err(GameError::NotSelectingCharactersState));

pub fn round(&self) -> Result<&Round, GameError>

Tries to get a &Round state. Returns an error if the game is not in a round state.

Examples

let game = GameState::Lobby(Lobby::default());
assert_eq!(game.round(), Err(GameError::NotRoundState));

pub fn round_mut(&mut self) -> Result<&mut Round, GameError>

Tries to get a &mutRound state. Returns an error if the game is not in a round state.

Examples

let mut game = GameState::Lobby(Lobby::default());
assert_eq!(game.round_mut(), Err(GameError::NotRoundState));

pub fn bankertarget(&self) -> Result<&BankerTargetRound, GameError>

Tries to get a &Round state. Returns an error if the game is not in a round state.

Examples

let game = GameState::Lobby(Lobby::default());
assert_eq!(game.bankertarget(), Err(GameError::NotBankerTargetState));

pub fn bankertarget_mut(&mut self) -> Result<&mut BankerTargetRound, GameError>

Tries to get a &mutRound state. Returns an error if the game is not in a round state.

Examples

let mut game = GameState::Lobby(Lobby::default());
assert_eq!(game.round_mut(), Err(GameError::NotRoundState));

pub fn results(&self) -> Result<&Results, GameError>

Tries to get a &Results state. Returns an error if the game is not in a results state.

Examples

let game = GameState::Lobby(Lobby::default());
assert_eq!(game.results(), Err(GameError::NotResultsState));

pub fn results_mut(&mut self) -> Result<&mut Results, GameError>

Tries to get a &mutResults state. Returns an error if the game is not in a results state.

Examples

let mut game = GameState::Lobby(Lobby::default());
assert_eq!(game.results_mut(), Err(GameError::NotResultsState));

pub fn start_game<P: AsRef<Path>>( &mut self, data_path: P, ) -> Result<(), GameError>

Starts the game if enough players are in the lobby. If the lobby has between 4 and 7 players inclusive, turns the state from a Lobby into a SelectingCharacters. Takes in a path that should point to an instance of boardgame.json, which holds the information about what cards each deck should be filled with.

pub fn player_select_character( &mut self, id: PlayerId, character: Character, ) -> Result<(), GameError>

Allows a player with id to select character if that character is available. If this was the last player to select a character, the state will be transformed from SelectingCharacters to Round

pub fn end_player_turn(&mut self, id: PlayerId) -> Result<TurnEnded, GameError>

Allows player with id to end their turn. If it was the last player in a round, transforms the internal state from Round back to SelectingCharacters. If it was the last turn of the game, transforms the internal state from Round into Results