game::game

Struct Deck

pub struct Deck<T> {
    pub image_back_url: Arc<String>,
    pub deck: Vec<T>,
    /* private fields */
}

A wrapper struct around Vec<T> which allows for easy interaction with it as a deck of cards.

Fields
image_back_url: Arc<String>

The back url of the particular deck

deck: Vec<T>

The list of actual cards

Implementations

impl<T: Clone> Deck<T>

pub fn new(deck: Vec<T>) -> Self

Creates a new Deck<T> based on a Vec<T>

Examples

let deck = Deck::new(vec![1, 2, 3]);
assert_eq!(deck.deck, [1, 2, 3]);

pub fn new_with_url(deck: Vec<T>, url: &str) -> Self

Creates a new Deck<T> based on a Vec<T> and an url which should point to the back of the deckโs cards in the asset folder.

Examples

let url = "assets/cards/card_back.svg";

let deck = Deck::new_with_url(vec![1, 2, 3], url);
assert_eq!(deck.deck, [1, 2, 3]);
assert_eq!(deck.image_back_url.as_str(), url);

pub fn draw(&mut self) -> T

Draws a new card from the deck. If the deck ran out it is restored from the backup deck, reshuffled and then a card is drawn from that new deck instead.

impl<T> Deck<T>

pub fn len(&self) -> usize

Returns the number of elements in the deck, also referred to as its โlengthโ.

pub fn is_empty(&self) -> bool

Returns true if the deck contains no elements.

pub fn set_image_back_url(&mut self, url: &str)

Sets the card url of the back image of the cards in the deck.

pub fn put_back(&mut self, card: T)

Puts back a card on the bottom of the deck

pub fn shuffle(&mut self)

Randomly reshuffles the deck