game::game

Struct Round

pub struct Round { /* private fields */ }

State containing all information related to the round state of the game. In the round stage, players each play a turn where they can draw cards, play cards and use their character ability. After every player has played a turn, players will be able to select characters again. If one player reached six or more assets during a round, the game will move to Results instead.

Implementations

impl Round

pub fn player(&self, id: PlayerId) -> Result<&RoundPlayer, GameError>

Get a reference to a RoundPlayer based on a specific PlayerId. Note that the players are in order, so id 0 refers to the player at index 0 and so on. See Players::player for further information

pub fn player_mut( &mut self, id: PlayerId, ) -> Result<&mut RoundPlayer, GameError>

Get a mutable reference to a RoundPlayer based on a specific PlayerId. Note that the players are in order, so id 0 refers to the player at index 0 and so on. See Players::player_mut for further information

pub fn player_from_character( &self, character: Character, ) -> Option<&RoundPlayer>

Get a reference to a RoundPlayer based on a specific character. Note that the players are in order, so id 0 refers to the player at index 0 and so on.

pub fn player_by_name(&self, name: &str) -> Result<&RoundPlayer, GameError>

Get a reference to a RoundPlayer based on a specific name.

pub fn current_player(&self) -> &RoundPlayer

Get a reference to the RoundPlayer whose turn it is.

pub fn next_player(&self) -> Option<&RoundPlayer>

Get a reference to the RoundPlayer whose turn is up next. If the current player is the last player, returns None instead.

NOTE: this will exclude players who will be skipped this round for one reason or another.

pub fn next_player_mut(&mut self) -> Option<&mut RoundPlayer>

Get a mutable reference to the RoundPlayer whose turn is up next. If the current player is the last player, returns None instead.

NOTE: this will exclude players who will be skipped this round for one reason or another.

pub fn players(&self) -> &[RoundPlayer]

Gets a slice of all players in the lobby. See Players::players for further information

pub fn open_characters(&self) -> &[Character]

Gets a slice containing all characters that cannot be picked by anyone this round.

pub fn banker_target(&self) -> Option<Character>

Gets the character who is currently targeted by the banker if one is available

pub fn player_info(&self, id: PlayerId) -> Vec<PlayerInfo>

Gets the PlayerInfo for each player, excluding the player that has the same id as id.

pub fn current_market(&self) -> &Market

Gets the current market

pub fn is_final_round(&self) -> bool

Gets whether or not this is the final round

pub fn player_get_fireble_characters(&mut self) -> Vec<Character>

Gets a list of characters that are available to be fired this round. This will exclude the list of Round::open_characters as well as characters that have already been skipped or fired this round.

pub fn player_get_regulator_swap_players(&mut self) -> Vec<RegulatorSwapPlayer>

Gets the number of assets and liabilities for each player the regulator can choose to swap with. This excludes their own cards.

pub fn player_play_card( &mut self, id: PlayerId, card_idx: usize, ) -> Result<PlayerPlayedCard, GameError>

Allows player with id id to play a card from their hand at index card_idx. If this player was the first to buy their first, second, third, fourth, fifth, seventh, eight or ninth asset, a new market and corresponding triggered events will be returned. The card that was played will also be returned.

pub fn player_redeem_liability( &mut self, id: PlayerId, liability_idx: usize, ) -> Result<(), GameError>

This allows player with id id to redeem a liability at index liability_idx if they are the CFO and if they can afford to pay off the debt. If they can redeem the liability, it will be added back into the deck.

pub fn player_draw_card( &mut self, id: PlayerId, card_type: CardType, ) -> Result<Either<&Asset, &Liability>, GameError>

This allows player with id id to draw a card of card type card_type. If they were allowed to draw that card, a reference to the card will be returned.

pub fn player_give_back_card( &mut self, id: PlayerId, card_idx: usize, ) -> Result<CardType, GameError>

This allows player with id id to give back a card from their hand at index card_idx. If they were able to give back the card, the card type of this card will be returned.

pub fn player_get_bonus_cash_character( &mut self, player_id: PlayerId, ) -> Result<u8, GameError>

Gets players bonus cash based on their characters color. If successfull returns gold.

pub fn player_fire_character( &mut self, id: PlayerId, character: Character, ) -> Result<Character, GameError>

This allows player with id id to fire a player who has character character if they are the shareholder. If this is successful, the player who got fired will not play their turn this round.

pub fn player_terminate_credit_character( &mut self, id: PlayerId, character: Character, ) -> Result<Character, GameError>

This allows player with id id to fire a player who has character character if they are the shareholder. If this is successful, the player who got fired will not play their turn this round.

pub fn player_swap_with_deck( &mut self, id: PlayerId, card_idxs: Vec<usize>, ) -> Result<AssetLiabilityCount, GameError>

This allows player with id id to swap a list of cards from their hand at indexes card_idxs with the deck. If succesful, this function returns the number of cards that were swapped with the deck in total.

pub fn player_swap_with_player( &mut self, id: PlayerId, target_id: PlayerId, ) -> Result<HandsAfterSwap, GameError>

This allows a player with id id to swap their hand of cards with a player with id target_id. If succesful, a copy of each player’s new hand is returned.

pub fn player_divest_asset( &mut self, id: PlayerId, target_id: PlayerId, asset_idx: usize, ) -> Result<u8, GameError>

This allows a player with id id to force player with id target_id to divest an asset at index asset_idx for market value minus 1. If succesful, returns the amount of gold it cost to divest the asset for.

pub fn get_divest_assets( &mut self, id: PlayerId, ) -> Result<Vec<DivestPlayer>, GameError>

Gets a list of DivestPlayer, which contains their player id as well as each asset that can be divested as well as the current cost to do so. This list excludes their own cards.

pub fn skipped_characters(&self) -> Vec<Character>

Gets a list of characters that are skipped between the turns of two players. Characters are called in order, so if any character is called but unavailable for any reason (not selected, fired or otherwise skipped), they will be added to this list.

pub fn leave(&mut self, id: PlayerId) -> Result<(), GameError>

Sets a player as disconnected

pub fn rejoin(&mut self, id: PlayerId) -> Result<&RoundPlayer, GameError>

Allows a player to rejoin

pub fn played_characters(&self) -> Vec<(PlayerId, Character)>

Returns a list of all players that have already taken their turn this round and their characters