Struct game::info::Details

pub struct Details {
    pub draw_extra: ToggleZoomedBuilder,
    pub tooltips: Vec<(Polygon, Text, (TripID, Time))>,
    pub hyperlinks: HashMap<String, Tab>,
    pub warpers: HashMap<String, ID>,
    pub time_warpers: HashMap<String, (TripID, Time)>,
    pub can_jump_to_time: bool,
}

Fields
draw_extra: ToggleZoomedBuilder

Draw extra things when unzoomed or zoomed.

tooltips: Vec<(Polygon, Text, (TripID, Time))>

Show these tooltips over the map. If the tooltip is clicked, time-warp and open the info panel.

hyperlinks: HashMap<String, Tab>

When a button with this label is clicked, open this info panel tab instead.

warpers: HashMap<String, ID>

When a button with this label is clicked, warp to this ID.

time_warpers: HashMap<String, (TripID, Time)>

When a button with this label is clicked, time-warp and open the info panel for this trip.

can_jump_to_time: bool