Module game::info::trip

Structs

OpenTrip

Functions

cancelled

endpoint

finished

future

highlight_slow_intersections
Highlights intersections which were "slow" on the map

highlight_slow_lanes
Highlights lanes which were "slow" on the map

make_elevation

make_timeline
Draws the timeline for a single trip, with tooltips

make_trip_details
Creates the timeline, location warp, and time warp buttons for one trip, and draws the route on the map.

ongoing