Function game::info::trip::make_trip_details

fn make_trip_details(
    ctx: &mut EventCtx<'_>, 
    app: &App, 
    trip_id: TripID, 
    open_trip: &mut OpenTrip, 
    details: &mut Details, 
    phases: Vec<TripPhase>, 
    map_for_pathfinding: &Map, 
    progress_along_path: Option<f64>
) -> Widget

Creates the timeline, location warp, and time warp buttons for one trip, and draws the route on the map.