Module game::ltn

Re-exports

pub use browse::BrowseNeighborhoods;

pub use filters::DiagonalFilter;

pub use filters::ModalFilters;

pub use partition::NeighborhoodID;

pub use partition::Partitioning;

Modules

auto

Experiments to make a neighborhood be low-traffic by automatically placing filters to prevent all rat runs.

browse

connectivity

draw_cells

export

filters

partition

pathfinding

per_neighborhood

rat_run_viewer

rat_runs

select_boundary

Structs

Cell

A partitioning of the interior of a neighborhood based on driving connectivity

DistanceInterval

An interval along a road’s length, with start < end.

Neighborhood

Functions

find_cells

floodfill