game

Module player

This file contains all four player states, as well as functionality for those players and ways to interact with them.

Structs

Asset

Representation of an asset card. Each asset has a gold and a silver value, as well as an associated color. Some cards alse have an AssetPowerup.

BankerTargetPlayer

The player type that corresponds to the BankerTargetRound stage of the game.

ChangeAssetColorData

A type that represents the changes made with the AssetPowerup::CountAsAnyColor asset ability. It contains the index of the asset that was changed, as well as its original color.

DivestAsset

Represents an asset that can be divested from a certain player including the cost of doing so.

DivestPlayer

Utility struct used to represent each asset that can be divested from a player including the cost of doing so.

Liability

Representation of a liability card. Each liability has an associated gold value as well as a LiabilityType, which determines how expensive it is to issue this liability.

LobbyPlayer

The player type corresponding to the Lobby state of the game.

PayBankerPlayer

Utility struct that represents a player confirming their selection of assets and liabilities to issue to pay the banker.

PlayableAssets

a representation of how many assets of each color a certain player is allowed to buy this round.

PlayerId

A wrapper around u8 which represents a player’s id.

PlayerInfo

Publicly available information for each player. This contains information that you would be able to see from another player if you were looking at what they have on the table. You cannot see their hand, but you can see the amount of asset cards and liability cards they have.

RegulatorSwapPlayer

Utility struct used to represent the amount of asset cards and liability cards a certain player has.

ResultsPlayer

The player type that corresponds to the Results stage of the game. During the results stage, each player can calculate and see their score.

RoundPlayer

The player type that corresponds to the Round stage of the game. During the round stage, each player has selected a character.

SelectingCharactersPlayer

The player type that corresponds to the SelectingCharacters stage of the game. In this stage, players may not have a selected character yet.

SilverIntoGoldData

A type that represents the changes made with the AssetPowerup::SilverIntoGold asset ability. It contains the index of the asset that was changed, as well as its original silver value.

ToggleChangeAssetColor

The representation of the result of toggling with AssetPowerup::CountAsAnyColor.

ToggleSilverIntoGold

The representation of the result of toggling with AssetPowerup::SilverIntoGold.

Enums

AssetPowerup

A certain powerup some assets have. These specify special actions this asset allows a player to take at the end of the game.

CardType

A card type used in relation to actions taken with player’s hands. Can either be Asset or Liability.

Character

An enum containing all characters currently in the game in the order in which they are called.

Color

Represtation of the colors associated with all assets as well as some selectable characters.

LiabilityType

The liability type determines the cost of lending for that particular liability.

Traits

GetPlayerInfo

Trait that should be implemented for each player type to be able to transform its internal data into publicly displayable PlayerInfo.