game::player

Struct PlayableAssets

pub struct PlayableAssets {
    total: u8,
    red_cost: u8,
    green_cost: u8,
    purple_cost: u8,
    yellow_cost: u8,
    blue_cost: u8,
}

a representation of how many assets of each color a certain player is allowed to buy this round.

Fields
total: u8red_cost: u8green_cost: u8purple_cost: u8yellow_cost: u8blue_cost: u8
Implementations

impl PlayableAssets

pub fn total(&self) -> u8

The total unit value of assets a player can buy

pub fn color_cost(&self, color: Color) -> u8

The unit cost of buying an asset of a certain color.