game::player

Struct ResultsPlayer

pub struct ResultsPlayer { /* private fields */ }

The player type that corresponds to the Results stage of the game. During the results stage, each player can calculate and see their score.

Implementations

impl ResultsPlayer

pub fn new(player: RoundPlayer, market: &Market) -> Self

Creates a new ResultsPlayer based on a RoundPlayer and a given Market. Because of the MinusIntoPlus asset ability, each player keeps track of their own market.

pub fn id(&self) -> PlayerId

Gets the id of the player

pub fn name(&self) -> &str

Gets the name of the player

pub fn cash(&self) -> u8

Gets the amount of cash of the player

pub fn assets(&self) -> &[Asset]

Gets a list of bought assets of the player

pub fn liabilities(&self) -> &[Liability]

Gets a list of issued liabilities of the player

pub fn hand(&self) -> &[Either<Asset, Liability>]

Gets the hand of this player.

pub fn market(&self) -> &Market

Gets the player’s personal market.

pub fn toggle_minus_into_plus( &mut self, color: Color, ) -> Result<&Market, GameError>

Resets back to the passed final_market and then turns the minus of a certain color into a zero or a zero into a plus.

pub fn toggle_silver_into_gold( &mut self, asset_idx: usize, ) -> Result<ToggleSilverIntoGold, GameError>

Turns the silver value of one of this player’s assets into gold.

pub fn toggle_change_asset_color( &mut self, asset_idx: usize, color: Color, ) -> Result<ToggleChangeAssetColor, GameError>

Turns the color of one of this player’s assets into another color. If they already did this, Also resets that asset’s color.

pub fn confirm_asset_ability( &mut self, asset_idx: usize, ) -> Result<(), GameError>

Asset abilities are toggleable by default. This function confirms the current configuration, after which a player cannot toggle this particular index anymore.

pub fn all_five_colors_bonus(&self) -> u8

Returns 5 if this player has bought assets of each of the 5 colors, 0 otherwise.

pub fn six_assets_bonus(&self) -> u8

Returns 4 if this player was the first to reach 6 assets, even if they were forced to divest that asset afterwards. Returns 2 if this player owns 6 or more assets and was not the first to reach 6 assets. Returns 0 otherwise.

pub fn total_gold(&self) -> u8

Gets tho total gold value of all assets this player owns

pub fn total_silver(&self) -> u8

Gets tho total silver value of all assets this player owns

pub fn trade_credit(&self) -> u8

Gets the amount of trade credit debt this player has.

pub fn bank_loan(&self) -> u8

Gets the amount of bank loan debt this player has.

pub fn bonds(&self) -> u8

Gets the amount of bonds debt this player has.

pub fn color_value(&self, color: Color) -> f64

Gets the value of all assets of a certain color this player has

pub fn fcf(&self) -> f64

Calculates the fcf (total market value of all player’s assets) for this player.

pub fn score(&self) -> f64

Gets the final score for this player