game::player

Struct RoundPlayer

pub struct RoundPlayer {

    pub(super) id: PlayerId,
    pub(super) name: String,
    pub(super) cash: u8,
    pub(super) assets: Vec<Asset>,
    pub(super) liabilities: Vec<Liability>,
    pub(super) character: Character,
    pub(super) hand: Vec<Either<Asset, Liability>>,
    pub(super) cards_drawn: Vec<usize>,
    pub(super) bonus_draw_cards: u8,
    pub(super) assets_to_play: u8,
    pub(super) playable_assets: PlayableAssets,
    pub(super) liabilities_to_play: u8,
    pub(super) total_cards_drawn: u8,
    pub(super) total_cards_given_back: u8,
    pub(super) has_used_ability: bool,
    pub(super) has_gotten_bonus_cash: bool,
    pub(super) was_first_to_six_assets: bool,
    pub(super) is_human: bool,

}

The player type that corresponds to the Round stage of the game. During the round stage, each player has selected a character.

Fields
id: PlayerIdname: Stringcash: u8assets: Vec<Asset>liabilities: Vec<Liability>character: Characterhand: Vec<Either<Asset, Liability>>cards_drawn: Vec<usize>bonus_draw_cards: u8assets_to_play: u8playable_assets: PlayableAssetsliabilities_to_play: u8total_cards_drawn: u8total_cards_given_back: u8has_used_ability: boolhas_gotten_bonus_cash: boolwas_first_to_six_assets: boolis_human: bool
Implementations

impl RoundPlayer

pub fn id(&self) -> PlayerId

Gets the id of the player

pub fn name(&self) -> &str

Gets the name of the player

pub fn cash(&self) -> u8

Gets the amount of cash of the player

pub(crate) fn _set_cash(&mut self, cash: u8)

pub fn assets(&self) -> &[Asset]

Gets a list of bought assets of the player

pub fn liabilities(&self) -> &[Liability]

Gets a list of issued liabilities of the player

pub fn character(&self) -> Character

Gets the character for this player

pub fn hand(&self) -> &[Either<Asset, Liability>]

Gets the hand of this player.

pub(crate) fn enable_first_to_six_assets_bonus(&mut self)

The first player to get six assets gets a cash bonus of 2.

pub fn is_human(&self) -> bool

Gets the human state of this player

pub fn set_is_human(&mut self, human: bool)

Sets the human state of this player

pub fn has_used_ability(&self) -> bool

Returns true if the player has used their ability already

pub fn total_cards_drawn(&self) -> u8

Returns the amount of cards already drawn by the player

pub fn total_cards_given_back(&self) -> u8

Returns the amount of cards already given back by the player

pub fn cards_drawn(&self) -> &[usize]

Returns the list of drawn cards

fn update_cards_drawn(&mut self, card_idx: usize)

Adds a new card_idx to the list of cards drawn this round.

fn can_afford_asset(&self, asset: &Asset) -> bool

fn can_play_asset(&self, color: Color) -> bool

Checks whether or not a player can play an asset of a certain color.

pub fn can_play_liability(&self) -> bool

Checks whether or not this player can still issue a liability.

pub fn assets_to_play(&self) -> u8

Returns the budget for assets this player can still play.

pub fn liabilities_to_play(&self) -> u8

Returns the number of liabilities this player can still issue.

pub(crate) fn redeem_liability( &mut self, liability_idx: usize, ) -> Result<Liability, RedeemLiabilityError>

Redeems a liability for a player by paying for it in cash. If succesful, returns the liability that was redeemed.

pub fn fire_character( &mut self, character: Character, ) -> Result<Character, FireCharacterError>

Tries to fire a character. If succesful, returns that character.

pub fn terminate_credit( &mut self, character: Character, ) -> Result<Character, TerminateCreditCharacterError>

Tries to terminate credit line of character. If succesful, returns that character.

pub fn swap_with_deck( &mut self, card_idxs: Vec<usize>, asset_deck: &mut Deck<Asset>, liability_deck: &mut Deck<Liability>, ) -> Result<AssetLiabilityCount, SwapError>

Swaps a list of card indexes card_idxs with the deck. Each asset that is swapped is put back into the deck and each liability that is swapped is put back into the liability deck. If succesful, returns the total number of assets (left) and liabilties (right) that were removed from the player’s hand. After this action, the player is able to draw the total number of returned cards.

pub fn swap_with_player( &mut self, target: &mut RoundPlayer, ) -> Result<(), SwapError>

Tries to swap the hands of this player, if they are the Regulator, with the hands of the target player.

pub fn remove_asset( &mut self, asset_idx: usize, ) -> Result<Asset, DivestAssetError>

Removes an asset from this player at index asset_idx. If succesful, returns the asset that was removed from the player.

pub fn divest_asset( &mut self, player: &RoundPlayer, asset_idx: usize, market: &Market, ) -> Result<u8, DivestAssetError>

Checks if this player can divest an asset at index asset_idx from the target player. If they can, the cost of doing so is returned.

pub(crate) fn play_card( &mut self, card_idx: usize, ) -> Result<Either<Asset, Liability>, PlayCardError>

Plays card in players hand with index card_idx. If that index is valid and they are allowed to play that card, it is returned.

fn draw_card( &mut self, card: Either<Asset, Liability>, ) -> Either<&Asset, &Liability>

Makes the player draw a new card to their hand.

pub(crate) fn draw_asset( &mut self, deck: &mut Deck<Asset>, ) -> Result<&Asset, DrawCardError>

Draws a new asset from the deck, if they are allowed. If succesful, a reference to this asset is returned.

pub(crate) fn draw_liability( &mut self, deck: &mut Deck<Liability>, ) -> Result<&Liability, DrawCardError>

Draws a new liability from the deck, if they are allowed. If succesful, a reference to this liability is returned.

pub(crate) fn give_back_card( &mut self, card_idx: usize, ) -> Result<Either<Asset, Liability>, GiveBackCardError>

Makes this player give back a card from the cards they drew this round at index card_idx. If succesful, the card that was given back is returned.

pub fn should_give_back_cards(&self) -> bool

Checks whether or not this player should still give back cards.

pub fn can_draw_cards(&self) -> bool

Checks whether or not this player can still draw any more cards

pub fn draws_n_cards(&self) -> u8

Gets the number of cards this player can draw in total

pub fn gives_back_n_cards(&self) -> u8

Gets the number of cards this player should give back in total.

pub fn playable_assets(&self) -> PlayableAssets

Gets this player’s PlayableAssets, which is a representation of how many assets of each color this player can buy this round.

pub fn playable_liabilities(&self) -> u8

Gets the amount of liabilities this player can play this round.

pub fn turn_start_cash(&self) -> u8

Gets the amount of cash this player gets to start their turn.

pub fn asset_bonus(&self) -> i16

Gets the amount of cash this player gets based on the character they chose and the assets they own.

pub fn market_condition_bonus(&self, current_market: &Market) -> i16

Gets the amount of cash this player gets based on the character they chose and the market condition of the color of that character.

pub fn turn_cash(&self) -> u8

Gets the total amount of cash this player receives at the start of their turn.

pub fn get_bonus_cash_character( &mut self, current_market: &Market, ) -> Result<u8, GetBonusCashError>

Get bonus gold a player can get on their turn based on their characters color and their bought assets

pub(crate) fn start_turn(&mut self)

Starts this player’s turn by givinig them their turn gold.