game::player

Trait GetPlayerInfo

pub trait GetPlayerInfo {
    // Required method
    fn info(&self) -> PlayerInfo;
}

Trait that should be implemented for each player type to be able to transform its internal data into publicly displayable PlayerInfo.

Required Methods

fn info(&self) -> PlayerInfo

Gets the publicly available info of this particular player.