game::utility

Module serde_asset_liability

Includes a tagged wrapper around Either<Asset, Liability> to make the serialized json nicer to interact with.

Modules

value

Serde module containing an implementation of serialize and deserialize which can be used to serialize Either<Asset, Liability> to a nicer tagged representation.

vec

Module intended to be used with #[serde(with)] containing an implementation of serialize and deserialize which can be used to serialize Vec<Either<Asset, Liability>> to a nicer tagged representation.

Enums

EitherAssetLiability

A nicer tagged representation of Either<Asset, Liability> which looks much better when serialized.