game::utility::serde_asset_liability

Module vec

Module intended to be used with #[serde(with)] containing an implementation of serialize and deserialize which can be used to serialize Vec<Either<Asset, Liability>> to a nicer tagged representation.

Functions

deserialize

Deserializes the list of nicer tagged representations back into a Vec<Either<Asset, Liability>>

serialize

Serializes a Vec<Either<Asset, Liability>> into a nicer tagged representation