Module game::app

Structs

App
The top-level data that lasts through the entire game, no matter what state the game is in.

FindDelayedIntersections

Flags

PerMap
All of the state that’s bound to a specific map.

PerObjectActions

SessionState

ShowEverything

ShowLayers

Traits

ShowObject

Functions

add_study_area
Load an extra GeoJSON file, and add the area to the map dynamically.

Type Definitions

Transition