Struct game::app::App

pub struct App {
    pub primary: PerMap,
    pub secondary: Option<PerMap>,
    pub cs: ColorScheme,
    pub opts: Options,
    pub per_obj: PerObjectActions,
    pub session: SessionState,
}

The top-level data that lasts through the entire game, no matter what state the game is in.

Fields
primary: PerMapsecondary: Option<PerMap>

This is only used right now for a debug mode to compare two versions of the same map. In the future, it might be used for some kind of live “side-by-side” comparison.

cs: ColorSchemeopts: Optionsper_obj: PerObjectActionssession: SessionState

Static data that lasts the entire session. Use sparingly.

Implementations

impl App

pub fn has_prebaked(&self) -> Option<(&MapName, &String)>

pub fn prebaked(&self) -> &Analytics

pub fn set_prebaked(&mut self, prebaked: Option<(MapName, String, Analytics)>)

pub fn draw(
&self,
g: &mut GfxCtx<'_>,
opts: DrawOptions,
show_objs: &dyn ShowObject
)

pub fn recalculate_current_selection(&mut self, ctx: &EventCtx<'_>)

Assumes some defaults.

pub fn mouseover_unzoomed_roads_and_intersections(
&self,
ctx: &EventCtx<'_>
) -> Option<ID>

pub fn mouseover_unzoomed_intersections(&self, ctx: &EventCtx<'_>) -> Option<ID>

pub fn mouseover_unzoomed_buildings(&self, ctx: &EventCtx<'_>) -> Option<ID>

pub fn mouseover_unzoomed_everything(&self, ctx: &EventCtx<'_>) -> Option<ID>

pub fn mouseover_debug_mode(
&self,
ctx: &EventCtx<'_>,
show_objs: &dyn ShowObject
) -> Option<ID>

fn calculate_current_selection(
&self,
ctx: &EventCtx<'_>,
show_objs: &dyn ShowObject,
debug_mode: bool,
unzoomed_roads_and_intersections: bool,
unzoomed_buildings: bool
) -> Option<ID>

fn get_renderables_back_to_front<'a>(
&'a self,
bounds: Bounds,
prerender: &Prerender,
agents: &'a mut AgentCache,
show_objs: &dyn ShowObject
) -> Vec<&'a (dyn Renderable + 'a)>

pub fn clear_everything(&mut self, ctx: &mut EventCtx<'_>)

Ensure the map edits are blank, reset the simulation, and blank out prebaked results.

impl App

pub fn click_on_intersection<S: Into<String>>(
&mut self,
ctx: &mut EventCtx<'_>,
label: S
) -> Option<IntersectionID>

If an intersection was clicked, return its ID.