Module game::edit

Re-exports

pub use self::roads::RoadEditor;

pub use self::routes::RouteEditor;

pub use self::stop_signs::StopSignEditor;

pub use self::traffic_signals::TrafficSignalEditor;

pub use self::validate::check_blackholes;

pub use self::validate::check_sidewalk_connectivity;

Modules

multiple_roads
After a single road has been edited, these states let the changes be copied to all similar road segments. Note that only lane configuration is copied, not speed limit or access restrictions.

roads

routes

stop_signs

traffic_signals

validate

zones

Structs

ConfirmDiscard

EditMode

LoadEdits

SaveEdits

Functions

apply_map_edits

can_edit_lane

cmd_to_id

make_changelist

make_topcenter

maybe_edit_intersection

speed_limit_choices