Module game::edit::traffic_signals::gmns

Structs

Record

Snapper
Snaps a line to a vehicle movement across an intersection. It uses movement endpoints and a hint about turn type to match.

Functions

add_crosswalks

cardinal_direction

import
This imports timing.csv from https://github.com/asu-trans-ai-lab/Vol2Timing. It operates in a best-effort / permissive mode, skipping over mismatched movements and other problems and should still be considered experimental.

import_all

parse_linestring

parse_osm_ids