Trait game::game::State

pub trait State: Downcast {
    fn event(&mut self, ctx: &mut EventCtx<'_>, app: &mut App) -> Transition;

    fn draw(&self, g: &mut GfxCtx<'_>, app: &App);

    fn draw_baselayer(&self) -> DrawBaselayer { ... }

    fn on_destroy(&mut self, _: &mut EventCtx<'_>, _: &mut App) { ... }
}

Required methods

fn event(&mut self, ctx: &mut EventCtx<'_>, app: &mut App) -> Transition

fn draw(&self, g: &mut GfxCtx<'_>, app: &App)

Loading content...
Provided methods

fn draw_baselayer(&self) -> DrawBaselayer

fn on_destroy(&mut self, _: &mut EventCtx<'_>, _: &mut App)

Loading content...
Implementations

impl dyn State

pub fn grey_out_map(g: &mut GfxCtx<'_>, app: &App)

impl dyn State

pub fn is<__T: State>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: State>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T: State>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T: State>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T: State>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.