Trait game::sandbox::gameplay::GameplayState

pub trait GameplayState: Downcast {
    fn event(
        &mut self, 
        ctx: &mut EventCtx<'_>, 
        app: &mut App, 
        controls: &mut SandboxControls, 
        actions: &mut Actions
    ) -> Option<Transition<App>>;

    fn draw(&self, g: &mut GfxCtx<'_>, app: &App);

    fn recreate_panels(&mut self, ctx: &mut EventCtx<'_>, app: &App);

    fn on_destroy(&self, _: &mut App) { ... }

    fn can_move_canvas(&self) -> bool { ... }

    fn can_examine_objects(&self) -> bool { ... }

    fn has_common(&self) -> bool { ... }

    fn has_tool_panel(&self) -> bool { ... }

    fn has_time_panel(&self) -> bool { ... }

    fn has_minimap(&self) -> bool { ... }
}

Required methods

fn event(
&mut self,
ctx: &mut EventCtx<'_>,
app: &mut App,
controls: &mut SandboxControls,
actions: &mut Actions
) -> Option<Transition<App>>

fn draw(&self, g: &mut GfxCtx<'_>, app: &App)

fn recreate_panels(&mut self, ctx: &mut EventCtx<'_>, app: &App)

Provided methods

fn on_destroy(&self, _: &mut App)

fn can_move_canvas(&self) -> bool

fn can_examine_objects(&self) -> bool

fn has_common(&self) -> bool

fn has_tool_panel(&self) -> bool

fn has_time_panel(&self) -> bool

fn has_minimap(&self) -> bool

Implementations

impl dyn GameplayState

pub fn is<__T: GameplayState>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: GameplayState>(
self: Box<Self>
) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: GameplayState>(
self: Rc<Self>
) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: GameplayState>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: GameplayState>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.