Struct game::Args

pub(crate) struct Args {

    pub(crate) flags: Flags,
    pub(crate) start_with_edits: Option<String>,
    pub(crate) cam: Option<String>,
    pub(crate) start_time: Option<Duration>,
    pub(crate) diff_map: Option<String>,
    pub(crate) dump_raw_events: bool,
    pub(crate) scale_factor: Option<f64>,
    pub(crate) dev: bool,
    pub(crate) color_scheme: Option<ColorSchemeChoice>,
    pub(crate) minimal_controls: bool,
    pub(crate) prebake: bool,
    pub(crate) tutorial_intro: bool,
    pub(crate) challenges: bool,
    pub(crate) sandbox: bool,
    pub(crate) proposals: bool,
    pub(crate) ungap: bool,
    pub(crate) devtools: bool,
    pub(crate) load_kml: Option<String>,
    pub(crate) challenge: Option<String>,
    pub(crate) tutorial: Option<usize>,
    pub(crate) actdev: Option<String>,
    pub(crate) actdev_scenario: Option<String>,
    pub(crate) compare_counts: Option<Vec<String>>,

}

Fields
flags: Flagsstart_with_edits: Option<String>

Start with these map edits loaded. This should be the name of edits, not a full path.

cam: Option<String>

Initially position the camera here. The format is an OSM-style zoom/lat/lon string (https://wiki.openstreetmap.org/wiki/Browsing#Other_URL_tricks).

start_time: Option<Duration>

Start the simulation at this time

diff_map: Option<String>

Load the map at this path as a secondary debug map to compare to the main one

dump_raw_events: bool

Print raw widgetry events to the console for debugging

scale_factor: Option<f64>

Override the monitor’s auto-detected scale factor

dev: bool

Dev mode exposes experimental tools useful for debugging, but that’d likely confuse most players.

color_scheme: Option<ColorSchemeChoice>

The color scheme for map elements, agents, and the UI.

minimal_controls: bool

When making a screen recording, enable this option to hide some UI elements

prebake: bool

Run a configured set of simulations and record prebaked data.

tutorial_intro: bool

Start at the tutorial intro screen

challenges: bool

Start by listing gameplay challenges

sandbox: bool

Start in the simulation sandbox mode

proposals: bool

Start by showing community proposals

ungap: bool

Launch Ungap the Map, a bike network planning tool

devtools: bool

Start by listing internal developer tools

load_kml: Option<String>

Start by showing this KMl file in a debug viewer

challenge: Option<String>

Start playing a particular challenge

tutorial: Option<usize>

Start on a particular tutorial stage

actdev: Option<String>

Start in ActDev mode for a particular site name.

actdev_scenario: Option<String>

Start by showing an ActDev scenario. Either “base” or “go_active”.

compare_counts: Option<Vec<String>>

Start in a tool for comparing traffic counts